AP_HAL_ChibiOS: more changes for 32 bit servo mask

+++ libraries/AP_HAL_ChibiOS/RCOutput_serial.cpp
@@ -50,7 +50,7 @@ bool RCOutput::dshot_send_command(pwm_group& group, uint8_t command, uint8_t cha
     group.dshot_waiter = rcout_thread_ctx;
     bool bdshot_telem = false;
 #ifdef HAL_WITH_BIDIR_DSHOT
-    uint16_t active_channels = group.ch_mask & group.en_mask;
+    uint32_t active_channels = group.ch_mask & group.en_mask;
     // no need to get the input capture lock
     group.bdshot.enabled = false;
     if ((_bdshot.mask & active_channels) == active_channels) {
